Move SNS delete cart from createNewOrder to up updateStatus
