+++ .env
@@ -1,0 +1,9 @@
+AWS_NODEJS_CONNECTION_REUSE_ENABLED=1
+REGION=eu-central-1
+STAGE=dev
+QUEUE=https://sqs.eu-central-1.amazonaws.com/780844780884/messagesQueue
+S3ARN=arn:aws:sns:eu-central-1:780844780884:images
+ORDERS_TABLE=orders-table
+USER_POOL_ID=eu-central-1_eciEUvwzp
+IS_OFFLINE=true
+API_ENDPOINT=http://localhost:3000/
